# openapi/admin-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chaos Admin API",
    "version": "1.0.0",
    "description": "Passwordless administration API for merchant accounts and headless commerce stores."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/admin/v1/stores/{store_id}/api-keys": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        }
      ],
      "get": {
        "operationId": "listApiKeys",
        "tags": [
          "API keys"
        ],
        "summary": "List Store API keys without secret material",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "API key page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCollectionEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      },
      "post": {
        "operationId": "createApiKey",
        "tags": [
          "API keys"
        ],
        "summary": "Create a Store-scoped API key",
        "description": "The plaintext secret is returned exactly once. Replaying the Idempotency-Key returns a conflict instead of disclosing it again.",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKey"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created; copy the one-time secret now",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreatedEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable session token"
      }
    },
    "parameters": {
      "StoreId": {
        "name": "store_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the previous page.",
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON, query, content type, or payload size",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or credential invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated caller is not allowed to access the resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Idempotency or uniqueness conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Domain or request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ApiKeyClass": {
        "type": "string",
        "enum": [
          "publishable",
          "secret"
        ]
      },
      "ApiKeyMode": {
        "type": "string",
        "enum": [
          "test",
          "live"
        ]
      },
...
      "ApiKeyScope": {
        "type": "string",
        "enum": [
          "analytics:write",
          "catalog:read",
          "carts:write",
          "checkout:write",
          "orders:read",
          "customers:write",
          "mcp:tools"
        ]
      },
      "CreateApiKey": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "class",
          "mode",
          "scopes"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "class": {
            "$ref": "#/components/schemas/ApiKeyClass"
          },
          "mode": {
            "$ref": "#/components/schemas/ApiKeyMode"
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            }
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "key_identifier",
          "display_suffix",
          "class",
          "mode",
          "scopes",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key_identifier": {
            "type": "string",
            "minLength": 16,
            "maxLength": 16
          },
          "display_suffix": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4
          },
          "class": {
            "$ref": "#/components/schemas/ApiKeyClass"
          },
          "mode": {
            "$ref": "#/components/schemas/ApiKeyMode"
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "revoked_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiKeyCreated": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "key_identifier",
          "display_suffix",
          "class",
          "mode",
          "scopes",
          "secret"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key_identifier": {
            "type": "string",
            "minLength": 16,
            "maxLength": 16
          },
          "display_suffix": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4
          },
          "class": {
            "$ref": "#/components/schemas/ApiKeyClass"
          },
          "mode": {
            "$ref": "#/components/schemas/ApiKeyMode"
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            }
          },
          "secret": {
            "type": "string",
            "format": "password",
            "readOnly": true,
            "x-sensitive": true,
            "description": "One-time plaintext API key. It cannot be retrieved again."
          }
        }
      },
      "ApiKeyCreatedEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ApiKeyCreated"
          }
        }
      },
...
      "PageMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "next_cursor": {
            "type": "string"
          }
        }
      },
      "ResponseMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "page"
        ],
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
...
      "ApiKeyCollectionEnvelope": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "reason"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    }
  }
}